==> notebooks/02_Preparation_Données/06_CLC.Rmd <==
---
title: "Occupation du sol"
output: html_notebook
---

```{sql}
create temporary table clc18_extract as 
(select a.* 
from clc18 as a, f_1000 as b 
where st_intersects(a.geom, b.geom) and a.geom && b.geom)
```


```{sql}
drop table clc18_friches;

create table clc18_friches as (
select 
a.site_pk as site_pk, 
b.id as id_clc,
b.code_18,
st_intersection(a.geom, b.geom) as geom
from friches_t1000 as a, clc18 as b
where st_intersects(a.geom, b.geom) and a.geom && b.geom)
```
