==> notebooks/02_Preparation_Données/09_Chef-Lieux.Rmd <==
---
title: "Distance aux chef-lieux"
output: html_notebook
---

```{r}
library(sf)
library(tidyverse)
```

```{r}
friches_points <- st_read("../data/friches/02_preprocessing/friches_points.gpkg") %>% st_set_crs(4326) %>% st_transform(2154) %>% select(site_pk)
```

```{r}
f_chef_lieux <- st_read("../data/chef-lieux/CHEF_LIEU.shp") %>% st_set_crs(2154)
```


```{r}
library(FNN)
coords1 <- st_coordinates(f_chef_lieux)
coords2 <- st_coordinates(friches_points)
nn <- get.knnx(coords1,
               coords2, k = 1)
```


```{r}
friches_points$NOM_CHF <- f_chef_lieux$NOM_CHF[nn$nn.index]
friches_points$INSEE_COM <- f_chef_lieux$INSEE_COM[nn$nn.index]
friches_points$STATUT <- f_chef_lieux$STATUT[nn$nn.index]
friches_points$distance_chef_lieu <- nn$nn.dist
```

Export
```{r}
st_write(friches_points, "../outputs/friches_mairie.gpkg", delete_dsn = T)
```
