==> notebooks/02_Preparation_Données/10_Identifiants.Rmd <==
---
title: "Ajout des identifiants"
output: html_notebook
---

```{r}
library(sf)
library(tidyverse)
```

```{r}
friches_points <- st_read("../data/friches/friches_points.gpkg")
friches_surfaces <- st_read("../data/friches/friches_surfaces.gpkg")
```

```{r}
friches_points2 <- friches_points %>% filter(source_nom != "Site Basias ou Basol non vérifié par le Cerema") %>% st_set_crs(4326)

friches_surfaces2 <- friches_surfaces %>% filter(source_nom != "Site Basias ou Basol non vérifié par le Cerema") %>% st_set_crs(4326)
```

```{r}
w <- st_is_empty(friches_points2) %>% which
friches_points3 <- friches_points2[-w, ]
```

```{r}
w <- st_is_empty(friches_surfaces2) %>% which
friches_surfaces3 <- friches_surfaces2[-w, ]
```

```{r}
friches_points3$site_pk <-1:nrow(friches_points3)
friches_surfaces3$site_pk <-1:nrow(friches_surfaces3)
```

```{r}
st_write(friches_points3, "../data/friches/02_preprocessing/friches_points.gpkg")
st_write(friches_surfaces3, "../data/friches/02_preprocessing/friches_surfaces.gpkg")
```
