==> notebooks/02_Preparation_Données/11_Tampons.Rmd <==
---
title: "Tampon de 1 Km"
output: html_notebook
---

```{r}
library(sf)
library(tidyverse)
```

```{r}
friches_points_t1000 <- st_read("../data/friches/02_preprocessing/friches_points.gpkg") %>% st_set_crs(4326) %>% st_transform(2154) %>% st_buffer(1000) %>% dplyr::select(site_pk)
```

```{r}
st_write(friches_points_t1000, "../data/friches/02_preprocessing/friches_t1000.gpkg")
```
