==> notebooks/02_Preparation_Données/12_BB.Rmd <==
---
title: "Friche and Chips"
output: html_notebook
---

```{r}
library(sf)
library(tidyverse)
```

```{r}
friches_surfaces <- st_read("../../data/friches/02_preprocessing/friches_surfaces.gpkg") %>% st_transform(2154)
```

```{r}
MBR <- function(p) {
  # Analyze the convex hull edges     
  a <- chull(p)                                   # Indexes of extremal points
  a <- c(a, a[1])                                 # Close the loop
  e <- p[a[-1],] - p[a[-length(a)], ]             # Edge directions
  norms <- sqrt(rowSums(e^2))                     # Edge lengths
  v <- e / norms                                  # Unit edge directions
  w <- cbind(-v[,2], v[,1])                       # Normal directions to the edges

  # Find the MBR
  vertices <- p[a, ]                              # Convex hull vertices
  x <- apply(vertices %*% t(v), 2, range)         # Extremes along edges
  y <- apply(vertices %*% t(w), 2, range)         # Extremes normal to edges
  areas <- (y[1,]-y[2,])*(x[1,]-x[2,])            # Areas
  k <- which.min(areas)                           # Index of the best edge (smallest area)

  # Form a rectangle from the extremes of the best edge
  cbind(x[c(1,2,2,1,1),k], y[c(1,1,2,2,1),k]) %*% rbind(v[k,], w[k,])
}
```

On calcule le coefficient d'élongation qui a la longueur minimale, la longueur maximale et l'élongation
```{r}
out <- vector(mode = "list")
for(i in 1:nrow(friches_surfaces)) {
  # print(i)
  coords <- friches_surfaces[i, ] %>% st_coordinates %>% {.[, c(1,2)]}
  bb <- MBR(coords)
  pol <- st_polygon(list(bb)) %>% st_sfc
  l <- st_cast(pol, "LINESTRING") %>% st_geometry
  l_int <- st_intersection(l, l)
  v <- sapply(1:4, function(x) {
    pts <- l_int[[1]][[x]]
    pt1 <- pts[1, ]
    pt2 <- pts[2, ]
    d <- (pt2[1] - pt1[1])^2 + (pt2[2] - pt1[2])^2
    return(d)})
  min_longueur <- min(v)
  max_longueur <- max(v)
  elongation <- min(v) / max(v)
  df <- data.frame(min_longueur, 
                   max_longueur, 
                   elongation)
  # st_geometry(df) <- l
  out[[i]] <- df
}
```

```{r}
res <- do.call(rbind, out)
```

```{r}
friches_surfaces2 <- cbind(friches_surfaces %>% select(site_pk), res)
```

```{r}
st_write(friches_surfaces2, "../../outputs/friches_and_chips.gpkg")
```





cbind()
